@@ -1,7 +1,7 @@
 package com.maodq.demo.internal;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.TypedValue;
 import android.view.Window;
 import android.widget.LinearLayout;
