@@ -1,7 +1,7 @@
 package com.maodq.demo.internal
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import android.widget.Button
 import com.maodq.demo.R
