@@ -1,8 +1,8 @@
 package com.maodq.demo.internal
 
 import android.content.Intent
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
@@ -10,16 +10,16 @@ import com.maodq.demo.base.BActivity
 import com.maodq.demo.base.BPresenter
 
 class MainPresenter : BPresenter() {
-    fun initView(rvList: RecyclerView, items: MutableList<Item>) {
+    fun initView(rvList: androidx.recyclerview.widget.RecyclerView, items: MutableList<Item>) {
         val linearLayoutManager =
-                LinearLayoutManager(rvList.context, LinearLayoutManager.VERTICAL, false)
+                androidx.recyclerview.widget.LinearLayoutManager(rvList.context, androidx.recyclerview.widget.LinearLayoutManager.VERTICAL, false)
         rvList.layoutManager = linearLayoutManager
         rvList.adapter = MyAdapter(items)
     }
 
     class MyAdapter(private val items: MutableList<Item>) :
-            RecyclerView.Adapter<RecyclerView.ViewHolder>() {
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+            androidx.recyclerview.widget.RecyclerView.Adapter<androidx.recyclerview.widget.RecyclerView.ViewHolder>() {
+        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): androidx.recyclerview.widget.RecyclerView.ViewHolder {
             val button = Button(parent.context)
             val density = parent.context.resources.displayMetrics.density
             button.layoutParams = ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, (density * 50).toInt())
@@ -38,12 +38,12 @@ class MainPresenter : BPresenter() {
             return items.size
         }
 
-        override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+        override fun onBindViewHolder(holder: androidx.recyclerview.widget.RecyclerView.ViewHolder, position: Int) {
             val item = items[position]
             val button = holder.itemView as Button
             button.text = item.name
         }
 
-        class MyViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
+        class MyViewHolder(itemView: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(itemView)
     }
 }
