@@ -1,7 +1,7 @@
 package com.maodq.demo.widget;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 
@@ -10,7 +10,7 @@ import com.maodq.demo.util.Util;
 /**
  * for Android event delivery
  */
-public class MyTextView extends android.support.v7.widget.AppCompatTextView {
+public class MyTextView extends androidx.appcompat.widget.AppCompatTextView {
     private static final String TAG = "TextView";
 
     public MyTextView(Context context) {
